+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Upgrade the `org.realityforge.javax.annotation` artifact to version `1.0.1`.
 * Upgrade the `org.realityforge.com.google.elemental2` artifacts to version `2.25`.
 * Upgrade the `au.com.stocksoftware.idea.codestyle` artifact to version `1.14`.
-* Upgrade the `org.realityforge.braincheck` artifact to version `1.24.0`.
+* Upgrade the `org.realityforge.braincheck` artifact to version `1.25.0`.
 * Add `JDepend` tests that enforce dependencies between packages.
 * Use capabilities in `braincheck` to start recording and verifying all invariant failures in `diagnostic_messages.json`.