+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-### Unreleased
+### [v0.10](https://github.com/spritz/spritz/tree/v0.10) (2020-01-20) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.09...v0.10)
+
+Changes in this release:
 
 * Upgrade the `com.google.guava` artifact to version `27.1-jre`.
 * Upgrade the `com.google.truth` artifact to version `0.44`.